The last version of Chute is from 2021 https://web.archive.org/web/20210119232116/https://www.getchute.com/
It appears that it is discontinued. Also, the company entry of
ESW Capital does not seem to have enough value to keep it as "archived".
It looks like a pure investment company, so there should be little
historical significance to keep it around in the TrackerDB data set.

fixes https://github.com/ghostery/trackerdb/issues/362