Update google_tag_manager.eno
Fine tune of filters as currently it is too blunt, wrongly classifying marketing pixels from Google Tags as "essential" from GTM.

e.g. https://www.googletagmanager.com/a?id=AW-10952806841&v=3..... is a Google Tag for Ads.	

Same set site used as for PR #406.

Note: Regex would be better here as the id param may not always be the first. However that appears to be an edge case so far.