Update google_adsense.eno (#666)
Google is using the domain googlesyndication.com for many purposes related to Ads, not just Adsense. For example, this filter is flagging the use of Adsense when we know for certain it is not in use on a site.

I have therefore removed all googlesyndication.com filters. Instead these domains will be picked up by the google_syndication.eno pattern. Both are category = advertising, so this feels like a cleaner and more robust approach.